Move virtual desktop related helper functions to separate files.
