+++ src/modules/fancyzones/lib/VirtualDesktopUtils.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "ZoneWindow.h"
+
+namespace VirtualDesktopUtils
+{
+    bool GetWindowDesktopId(HWND topLevelWindow, GUID* desktopId);
+    bool GetZoneWindowDesktopId(IZoneWindow* zoneWindow, GUID* desktopId);
+}
