+++ src/modules/fancyzones/lib/ZoneWindow.h
@@ -9,9 +9,6 @@ namespace ZoneWindowUtils
     const std::wstring& GetCustomZoneSetsTmpPath();
 
     std::wstring GenerateUniqueId(HMONITOR monitor, PCWSTR deviceId, PCWSTR virtualDesktopId);
-
-    bool GetWindowDesktopId(HWND topLevelWindow, GUID* desktopId);
-    bool GetZoneWindowDesktopId(IZoneWindow* zoneWindow, GUID* desktopId);
 }
 
 /**
